+++ component/component.go
@@ -82,7 +82,7 @@ type Host interface {
 	// Typically is used to find an extension by type or by full config name. Both cases
 	// can be done by iterating the returned map. There are typically very few extensions
 	// so there there is no performance implications due to iteration.
-	GetExtensions() map[configmodels.Extension]ServiceExtension
+	GetExtensions() map[configmodels.NamedEntity]Extension
 
 	// Return map of exporters. Only enabled and created exporters will be returned.
 	// Typically is used to find exporters by type or by full config name. Both cases
@@ -92,7 +92,7 @@ type Host interface {
 	// Note that an exporter with the same name may be attached to multiple pipelines and
 	// thus we may have an instance of the exporter for multiple data types.
 	// This is an experimental function that may change or even be removed completely.
-	GetExporters() map[configmodels.DataType]map[configmodels.Exporter]Exporter
+	GetExporters() map[configmodels.DataType]map[configmodels.NamedEntity]Exporter
 }
 
 // Factory interface must be implemented by all component factories.
@@ -119,19 +119,3 @@ type ConfigUnmarshaler interface {
 // intoCfg interface{}
 //   An empty interface wrapping a pointer to the config struct to unmarshal into.
 type CustomUnmarshaler func(componentViperSection *viper.Viper, intoCfg interface{}) error
-
-// ApplicationStartInfo is the information that is logged at the application start and
-// passed into each component. This information can be overridden in custom builds.
-type ApplicationStartInfo struct {
-	// Executable file name, e.g. "otelcol".
-	ExeName string
-
-	// Long name, used e.g. in the logs.
-	LongName string
-
-	// Version string.
-	Version string
-
-	// Git hash of the source code.
-	GitHash string
-}