+++ component/componenttest/error_waiting_host.go
@@ -61,10 +61,10 @@ func (ews *ErrorWaitingHost) GetFactory(_ component.Kind, _ configmodels.Type) c
 	return nil
 }
 
-func (ews *ErrorWaitingHost) GetExtensions() map[configmodels.Extension]component.ServiceExtension {
+func (ews *ErrorWaitingHost) GetExtensions() map[configmodels.NamedEntity]component.Extension {
 	return nil
 }
 
-func (ews *ErrorWaitingHost) GetExporters() map[configmodels.DataType]map[configmodels.Exporter]component.Exporter {
+func (ews *ErrorWaitingHost) GetExporters() map[configmodels.DataType]map[configmodels.NamedEntity]component.Exporter {
 	return nil
 }